Move zone getter/setter to mio settings service
