+++ libs/ts/movein/ngx/src/lib/modules/builder/modules/zone-select/zone-select.component.html
@@ -1,51 +1,53 @@
 <div class="container12">
   <p class="center component-instructions">Select your residence hall.</p>
 
-  <div class="zone-container">
-    <h3>North Side</h3>
-    <div class="hall-item-container">
-      <div *ngFor="let hall of zones.northSide.halls; let i = index;" (click)="saveResidence(hall)" (keyup.space)="saveResidence(hall)" (keyup.enter)="saveResidence(hall)" tabindex="0">
-        <div class="name">
-          <p [ngClass]="{'active': savedResidence && savedResidence.name === hall.name}">{{hall.name}}</p>
-        </div>
+  <ng-container *ngIf="savedResidence | async as residence">
+    <div class="zone-container">
+      <h3>North Side</h3>
+      <div class="hall-item-container">
+        <div *ngFor="let hall of zones.northSide.halls; let i = index;" (click)="saveResidence(hall)" (keyup.space)="saveResidence(hall)" (keyup.enter)="saveResidence(hall)" tabindex="0">
+          <div class="name">
+            <p [ngClass]="{'active': residence && residence.name === hall.name}">{{hall.name}}</p>
+          </div>
 
-        <div class="action" [ngClass]="{'active': savedResidence && savedResidence.name === hall.name}">
-          <p class="material-icons" *ngIf="savedResidence && savedResidence.name === hall.name">check</p>
-          <p class="material-icons" *ngIf="!savedResidence || savedResidence.name !== hall.name">arrow_forward</p>
+          <div class="action" [ngClass]="{'active': residence && residence.name === hall.name}">
+            <p class="material-icons" *ngIf="residence && residence.name === hall.name">check</p>
+            <p class="material-icons" *ngIf="!residence || residence.name !== hall.name">arrow_forward</p>
+          </div>
         </div>
       </div>
     </div>
-  </div>
 
-  <div class="zone-container">
-    <h3>South Side</h3>
-    <div class="hall-item-container">
-      <div *ngFor="let hall of zones.southSide.halls; let i = index;" (click)="saveResidence(hall)" (keyup.space)="saveResidence(hall)" (keyup.enter)="saveResidence(hall)" tabindex="0">
-        <div class="name">
-          <p [ngClass]="{'active': savedResidence && savedResidence.name === hall.name}">{{hall.name}}</p>
-        </div>
+    <div class="zone-container">
+      <h3>South Side</h3>
+      <div class="hall-item-container">
+        <div *ngFor="let hall of zones.southSide.halls; let i = index;" (click)="saveResidence(hall)" (keyup.space)="saveResidence(hall)" (keyup.enter)="saveResidence(hall)" tabindex="0">
+          <div class="name">
+            <p [ngClass]="{'active': residence && residence.name === hall.name}">{{hall.name}}</p>
+          </div>
 
-        <div class="action" [ngClass]="{'active': savedResidence && savedResidence.name === hall.name}">
-          <p class="material-icons" *ngIf="savedResidence && savedResidence.name === hall.name">check</p>
-          <p class="material-icons" *ngIf="!savedResidence || savedResidence.name !== hall.name">arrow_forward</p>
-        </div>
-      </div>
-    </div>
-  </div>
-
-  <div class="zone-container">
-    <h3>White Creek</h3>
-    <div class="hall-item-container">
-      <div *ngFor="let hall of zones.whiteCreek.halls; let i = index;" (click)="saveResidence(hall)" (keyup.space)="saveResidence(hall)" (keyup.enter)="saveResidence(hall)" tabindex="0">
-        <div class="name">
-          <p [ngClass]="{'active': savedResidence && savedResidence.name === hall.name}">{{hall.name}}</p>
-        </div>
-
-        <div class="action" [ngClass]="{'active': savedResidence && savedResidence.name === hall.name}">
-          <p class="material-icons" *ngIf="savedResidence && savedResidence.name === hall.name">check</p>
-          <p class="material-icons" *ngIf="!savedResidence || savedResidence.name !== hall.name">arrow_forward</p>
+          <div class="action" [ngClass]="{'active': residence && residence.name === hall.name}">
+            <p class="material-icons" *ngIf="residence && residence.name === hall.name">check</p>
+            <p class="material-icons" *ngIf="!residence || residence.name !== hall.name">arrow_forward</p>
+          </div>
         </div>
       </div>
     </div>
-  </div>
+
+    <div class="zone-container">
+      <h3>White Creek</h3>
+      <div class="hall-item-container">
+        <div *ngFor="let hall of zones.whiteCreek.halls; let i = index;" (click)="saveResidence(hall)" (keyup.space)="saveResidence(hall)" (keyup.enter)="saveResidence(hall)" tabindex="0">
+          <div class="name">
+            <p [ngClass]="{'active': residence && residence.name === hall.name}">{{hall.name}}</p>
+          </div>
+
+          <div class="action" [ngClass]="{'active': residence && residence.name === hall.name}">
+            <p class="material-icons" *ngIf="residence && residence.name === hall.name">check</p>
+            <p class="material-icons" *ngIf="!residence || residence.name !== hall.name">arrow_forward</p>
+          </div>
+        </div>
+      </div>
+    </div>
+  </ng-container>
 </div>
