+++ libs/ts/movein/ngx/src/lib/modules/map/services/move-in-out-settings/move-in-out-settings.service.ts
@@ -3,7 +3,8 @@ import { Injectable } from '@angular/core';
 import { EnvironmentService } from '@tamu-gisc/common/ngx/environment';
 import { LocalStoreService } from '@tamu-gisc/common/ngx/local-store';
 
-import { MoveDates, MoveInSettings } from '../../../../interfaces/move-in-out.interface';
+import { MoveDates, MoveInSettings, ResidenceHall, ResidenceZones } from '../../../../interfaces/move-in-out.interface';
+import { RESIDENCES } from '../../../../dictionaries/move-in-out.dictionary';
 
 @Injectable({
   providedIn: 'root'
@@ -12,6 +13,7 @@ export class MoveInOutSettingsService {
   private _settingsPrimaryKey = 'aggiemap-movein';
 
   public days: MoveDates = this.env.value('MoveInOutDates', false);
+  public zones: ResidenceZones = RESIDENCES;
 
   public get settings(): MoveInSettings {
     return this.store.getStorage({ primaryKey: this._settingsPrimaryKey });
@@ -44,6 +46,38 @@ export class MoveInOutSettingsService {
     });
 
     return confirm;
+  }
+
+  public saveResidence(res: ResidenceHall) {
+    // Iterate through residence hall zones and find the zone that includes the provided residence hall.
+    // This is used to determine which zone the residence hall belongs to.
+    const zone = Object.keys(this.zones).find((key) => {
+      return this.zones[key].halls.find((hall) => {
+        return hall.name === res.name;
+      });
+    });
+
+    if (zone !== undefined) {
+      // Append zone to the residence. Used to render map elements based on region
+      res.zone = this.zones[zone].name;
+
+      this.store.setStorageObjectKeyValue({
+        primaryKey: this._settingsPrimaryKey,
+        subKey: 'residence',
+        value: res
+      });
+
+      // Verify that the value store was successful.
+      const confirm = this.store.getStorageObjectKeyValue<ResidenceHall>({
+        primaryKey: this._settingsPrimaryKey,
+        subKey: 'residence'
+      });
+
+      return confirm;
+    } else {
+      console.error('Zone not found for hall ', res.name);
+      return null;
+    }
   }
 }
